Viash 0.5.11

+++ src/viash/viash_build/script.sh
@@ -39,7 +39,7 @@ command_builder+=(
 if [ "$par_mode" == "development" ]; then
   if [ "$par_no_cache" == "true" ]; then
     setup_strat="build"
-  else 
+  else
     setup_strat="cachedbuild"
   fi
 elif [ "$par_mode" == "integration" ]; then
@@ -82,13 +82,13 @@ fi
 
 # process queries
 if [ ! -z "$par_query" ]; then
-  command_builder+=( "--query" "$par_query" )
-fi
-if [ ! -z "$par_query_namespace" ]; then
-  command_builder+=( "--query_name" "$par_query_namespace" )
+  command_builder+=("--query" "$par_query")
 fi
 if [ ! -z "$par_query_name" ]; then
-  command_builder+=( "--query_namespace" "$par_query_name" )
+  command_builder+=("--query_name" "$par_query_name")
+fi
+if [ ! -z "$par_query_namespace" ]; then
+  command_builder+=("--query_namespace" "$par_query_namespace")
 fi
 
 # process config mods
@@ -96,27 +96,34 @@ if [ ! -z "$par_config_mod" ]; then
   IFS=";"
   for var in $par_config_mod; do
     unset IFS
-    command_builder+=( "--config_mod" "$var" )
+    command_builder+=("--config_mod" "$var")
   done
+fi
+
+if [ ! -z "$par_nextflow_variant" ]; then
+  command_builder+=(
+    --config_mod "<preparse> .platforms[.type == 'nextflow'].variant := '$par_nextflow_variant'"
+  )
 fi
 
 if [ ! -z "$par_registry" ]; then
   command_builder+=(
     --config_mod ".platforms[.type == 'docker'].target_registry := '$par_registry'"
-    --config_mod ".platforms[.type == 'nextflow'].registry := '$par_registry'"
+    --config_mod ".platforms[.type == 'nextflow' && .variant == 'legacy'].registry := '$par_registry'"
   )
 fi
 
 if [ ! -z "$par_organization" ]; then
   command_builder+=(
     --config_mod ".platforms[.type == 'docker'].target_organization := '$par_organization'"
-    --config_mod ".platforms[.type == 'nextflow'].organization := '$par_organization'"
+    --config_mod ".platforms[.type == 'nextflow' && .variant == 'legacy'].organization := '$par_organization'"
   )
 fi
 
 if [ ! -z "$par_namespace_separator" ]; then
   command_builder+=(
-    --config_mod ".platforms[.type == 'docker' || .type == 'nextflow'].namespace_separator := '$par_namespace_separator'"
+    --config_mod ".platforms[.type == 'docker'].namespace_separator := '$par_namespace_separator'"
+    --config_mod ".platforms[.type == 'nextflow' && .variant == 'legacy'].namespace_separator := '$par_namespace_separator'"
   )
 fi
 
@@ -127,7 +134,7 @@ if [ ! -z "$par_target_image_source" ]; then
 fi
 
 if [ ! -z "$par_platform" ]; then
-  command_builder+=( --platform "$par_platform" )
+  command_builder+=(--platform "$par_platform")
 fi
 
 ################ RUN COMMAND ################
@@ -136,6 +143,6 @@ fi
 if [ -z "$par_log" ]; then
   "$par_viash" "${command_builder[@]}"
 else
-  rm "$par_log"
+  [ ! -f "$par_log" ] || rm "$par_log"
   "$par_viash" "${command_builder[@]}" > >(tee -a "$par_log") 2> >(tee -a "$par_log")
 fi
